Remove steps from slider

+++ app/src/main/java/com/codingwithmitch/giffit/MainViewModel.kt
@@ -185,7 +185,7 @@ class MainViewModel : ViewModel() {
                                 resizedGifUri = null,
                                 originalGifSize = gifSize,
                                 adjustedBytes = gifSize,
-                                sizePercentage = 100f,
+                                sizePercentage = 100,
                                 capturedBitmaps = it.capturedBitmaps
                             )
                         )
@@ -350,19 +350,19 @@ class MainViewModel : ViewModel() {
             _state.value = this.copy(
                 resizedGifUri = null,
                 adjustedBytes = originalGifSize,
-                sizePercentage = 100f
+                sizePercentage = 100
             )
         }
     }
 
-    fun updateAdjustedBytes(adjustedBytes: Float) {
+    fun updateAdjustedBytes(adjustedBytes: Int) {
         check(state.value is DisplayGif) { "updateAdjustedBytes: Invalid state: ${state.value}" }
         _state.value = (state.value as DisplayGif).copy(
-            adjustedBytes = adjustedBytes.toInt()
+            adjustedBytes = adjustedBytes
         )
     }
 
-    fun updateSizePercentage(sizePercentage: Float) {
+    fun updateSizePercentage(sizePercentage: Int) {
         check(state.value is DisplayGif) { "updateSizePercentage: Invalid state: ${state.value}" }
         _state.value = (state.value as DisplayGif).copy(
             sizePercentage = sizePercentage
